This was a restructuring to separate commands only related to NMEA output properties and general commands (in this case it was only the output type choice) into different files and directories
